src/requirements.txt
alabaster==0.7.12
click==7.0
django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
packaging==20.4
pandas==1.1.3
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
urllib3==1.25.11
